requirements.txt
ollama
SpeechRecognition
piper-tts
sounddevice
numpy
PyAudio